requirements.txt
chroma-hnswlib==0.7.5
chromadb==0.5.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
openai
pandas==2.1.4
sphinx
sphinx-rtd-theme
streamlit==1.38.0
python-magic
chardet
secure-smtplib
gspread
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
fpdf
pysqlite3-binary
requests